commit application

+++ know-heart/src/main/groovy/com/qubaopen/survey/service/self/SelfPersistentService.groovy
@@ -182,7 +182,9 @@ public class SelfPersistentService {
 				it.mapStatistics = mapStatistics
 			}
 		}
-		mapRecordRepository.deleteByMapStatistics(mapStatistics)
+		if (!mapStatistics) {
+			mapRecordRepository.deleteByMapStatistics(mapStatistics)
+		}
 		mapStatisticsRepository.save(mapStatistics)
 		mapRecordRepository.save(result)
 	}
